Introduce classes/text_processor.py
Unused; committing now mostly so all of you can see what I'm planning.
